RIGHTSAPI-23: Provide events in case of change in rights
Add a mechanism allowing to compute a diff between two set of security
rules. The diff is currently computed on normalized rules, so it will
only find updates related to rights on couples (subject,state), for
other kind of changes it will produce added and deleted diff.
This improvment has been done in order to be able to send those diff
rules as data with the RightUpdatedEvents.